+++ rust/scx_arena/scx_arena/Cargo.toml
@@ -13,4 +13,4 @@ simplelog = "0.12"
 scx_utils = { path = "../../scx_utils", version = "1.0.24" }
 
 [build-dependencies]
-scx_utils = { path = "../../scx_utils", version = "1.0.24" }
+scx_cargo = { path = "../../scx_cargo", version = "1.0.24" }
