+++ rust/scx_cargo/Cargo.toml
@@ -8,4 +8,24 @@ repository = "https://github.com/sched-ext/scx"
 description = "Build time utilities for sched_ext schedulers"
 
 [dependencies]
-scx_utils = { path = "../scx_utils", version = "1.0.24" }
+anyhow = "1.0.65"
+lazy_static = "1.5.0"
+sscanf = "0.4"
+version-compare = "0.1"
+glob = "0.3.2"
+tracing = "0.1"
+tracing-subscriber = "0.3.20"
+libbpf-cargo = "=0.26.0-beta.1"
+tar = "0.4"
+libbpf-rs = "=0.26.0-beta.1"
+bindgen = ">=0.69"
+
+[build-dependencies]
+ruzstd = "0.8.1"
+tar = "0.4"
+walkdir = "2.5"
+tempfile = "3.19.1"
+
+[dev-dependencies]
+regex = "1.11.2"
+tempfile = "3.19.1"
