+++ rust/scx_cargo/src/lib.rs
@@ -3,4 +3,8 @@
 // This software may be used and distributed according to the terms of the
 // GNU General Public License version 2.
 
-pub use scx_utils::BpfBuilder;
+mod clang_info;
+pub use clang_info::ClangInfo;
+
+mod bpf_builder;
+pub use bpf_builder::BpfBuilder;
