+++ rust/scx_cargo/vmlinux.tar.zst
@@ -1,0 +1,1 @@
+../scx_utils/vmlinux.tar.zst
