+++ rust/scx_rustland_core/Cargo.toml
@@ -13,7 +13,7 @@ plain = "0.2.3"
 libbpf-rs = "=0.26.0-beta.1"
 libc = "0.2.175"
 seccomp = "0.1"
-scx_utils = { path = "../scx_utils", version = "1.0.24" }
+scx_cargo = { path = "../scx_cargo", version = "1.0.24" }
 
 [lib]
 name = "scx_rustland_core"
