+++ rust/scx_utils/Cargo.toml
@@ -37,6 +37,8 @@ tracing-subscriber = "0.3.20"
 libc = "0.2.175"
 nix = { version = "0.30.1", features = ["resource"] }
 
+scx_cargo = { path = "../scx_cargo", version = "1.0.24", optional = true }
+
 [build-dependencies]
 anyhow = "1.0.65"
 bindgen = ">=0.69"
@@ -44,6 +46,7 @@ glob = "0.3.2"
 lazy_static = "1.5.0"
 libbpf-cargo = "=0.26.0-beta.1"
 ruzstd = "0.8.1"
+scx_cargo = { path = "../scx_cargo", version = "1.0.24" }
 sscanf = "0.4"
 tar = "0.4"
 tempfile = "3.19.1"
@@ -55,6 +58,7 @@ walkdir = "2.5"
 default = []
 gpu-topology = ["dep:nvml-wrapper", "dep:nvml-wrapper-sys"]
 autopower = ["dep:zbus"]
+deprecated-build-support = ["dep:scx_cargo"]
 
 [[example]]
 name = "mangolog"
